+++ linopy/__init__.py
@@ -12,6 +12,7 @@ __version__ = version("linopy")
 # Note: For intercepting multiplications between xarray dataarrays, Variables and Expressions
 # we need to extend their __mul__ functions with a quick special case
 import linopy.monkey_patch_xarray  # noqa: F401
+from linopy._active_gate import active_gate
 from linopy.alignment import align
 from linopy.config import options
 from linopy.constants import (
@@ -67,6 +68,7 @@ __all__ = (
     "SolverFeature",
     "Variable",
     "Variables",
+    "active_gate",
     "align",
     "available_solvers",
     "licensed_solvers",